Update lexer.hpp
I added the picture of emplex to discord. I made this for our lexer right now... there is some things wrong with it for sure but its a pretty good start. I also know he changed the way single characters get passed thru so i dont think the semicolon or assign actually work
had to make one so that the project.cpp file had a way of knowing what emplex:: was